Split array packer float handler into header and source files
This required a bit of extra attention for the build system.
